New code for Premium (#10)
* Added module for new method for premium appointments, replaces selenium code with new method
